git_utils/requirements.txt
python-gitlab==2.5.0
